improve the blueprint macro docs
